:art: Extract function for constructing portal URI
