+++ lib/editor-binding.js
@@ -3,7 +3,7 @@
 const path = require('path')
 const {Range, Emitter, Disposable, CompositeDisposable} = require('atom')
 const getPathWithNativeSeparators = require('./get-path-with-native-separators')
-const getEditorURI = require('./get-editor-uri')
+const {getEditorURI} = require('./uri-helpers')
 const {FollowState} = require('@atom/teletype-client')
 
 module.exports =
