--- lib/get-editor-uri.js
@@ -1,3 +1,0 @@
-module.exports = function getEditorURI (portalId, editorProxyId) {
-  return 'atom://teletype/portal/' + portalId + '/editor/' + editorProxyId
-}
