+++ lib/guest-portal-binding.js
@@ -4,7 +4,7 @@ const BufferBinding = require('./buffer-binding')
 const EditorBinding = require('./editor-binding')
 const SitePositionsComponent = require('./site-positions-component')
 const getPathWithNativeSeparators = require('./get-path-with-native-separators')
-const getEditorURI = require('./get-editor-uri')
+const {getEditorURI} = require('./uri-helpers')
 const NOOP = () => {}
 
 module.exports =
