+++ lib/teletype-package.js
@@ -5,6 +5,7 @@ const PortalStatusBarIndicator = require('./portal-status-bar-indicator')
 const AuthenticationProvider = require('./authentication-provider')
 const CredentialCache = require('./credential-cache')
 const TeletypeService = require('./teletype-service')
+const {getPortalURI} = require('./uri-helpers')
 
 module.exports =
 class TeletypePackage {
@@ -124,7 +125,7 @@ class TeletypePackage {
   async copyHostPortalURL () {
     const manager = await this.getPortalBindingManager()
     const hostPortalBinding = await manager.getHostPortalBinding()
-    const url = `atom://teletype/portal/${hostPortalBinding.portal.id}`
+    const url = getPortalURI(hostPortalBinding.portal.id)
     atom.clipboard.write(url)
   }
