+++ lib/uri-helpers.js
@@ -1,0 +1,9 @@
+function getPortalURI (portalId) {
+  return 'atom://teletype/portal/' + portalId
+}
+
+function getEditorURI (portalId, editorProxyId) {
+  return getPortalURI(portalId) + '/editor/' + editorProxyId
+}
+
+module.exports = {getEditorURI, getPortalURI}
